+++ src/main/kotlin/app/revanced/api/configuration/routes/ManagerRoute.kt
@@ -1,9 +1,8 @@
 package app.revanced.api.configuration.routes
 
+import app.revanced.api.configuration.ApiRelease
+import app.revanced.api.configuration.ApiReleaseVersion
 import app.revanced.api.configuration.installNotarizedRoute
-import app.revanced.api.configuration.schema.APIManagerAsset
-import app.revanced.api.configuration.schema.APIRelease
-import app.revanced.api.configuration.schema.APIReleaseVersion
 import app.revanced.api.configuration.services.ManagerService
 import io.bkbn.kompendium.core.metadata.GetInfo
 import io.ktor.http.*
@@ -14,18 +13,9 @@ import io.ktor.server.routing.*
 import org.koin.ktor.ext.get as koinGet
 
 internal fun Route.managerRoute() = route("manager") {
-    configure()
-
-    // TODO: Remove this deprecated route eventually.
-    route("latest") {
-        configure(deprecated = true)
-    }
-}
-
-private fun Route.configure(deprecated: Boolean = false) {
     val managerService = koinGet<ManagerService>()
 
-    installManagerRouteDocumentation(deprecated)
+    installManagerRouteDocumentation()
 
     rateLimit(RateLimitName("weak")) {
         get {
@@ -33,7 +23,7 @@ private fun Route.configure(deprecated: Boolean = false) {
         }
 
         route("version") {
-            installManagerVersionRouteDocumentation(deprecated)
+            installManagerVersionRouteDocumentation()
 
             get {
                 call.respond(managerService.latestVersion())
@@ -42,34 +32,32 @@ private fun Route.configure(deprecated: Boolean = false) {
     }
 }
 
-private fun Route.installManagerRouteDocumentation(deprecated: Boolean) = installNotarizedRoute {
+private fun Route.installManagerRouteDocumentation() = installNotarizedRoute {
     tags = setOf("Manager")
 
     get = GetInfo.builder {
-        if (deprecated) isDeprecated()
         description("Get the current manager release")
         summary("Get current manager release")
         response {
             description("The latest manager release")
             mediaTypes("application/json")
             responseCode(HttpStatusCode.OK)
-            responseType<APIRelease<APIManagerAsset>>()
+            responseType<ApiRelease>()
         }
     }
 }
 
-private fun Route.installManagerVersionRouteDocumentation(deprecated: Boolean) = installNotarizedRoute {
+private fun Route.installManagerVersionRouteDocumentation() = installNotarizedRoute {
     tags = setOf("Manager")
 
     get = GetInfo.builder {
-        if (deprecated) isDeprecated()
         description("Get the current manager release version")
         summary("Get current manager release version")
         response {
             description("The current manager release version")
             mediaTypes("application/json")
             responseCode(HttpStatusCode.OK)
-            responseType<APIReleaseVersion>()
+            responseType<ApiReleaseVersion>()
         }
     }
 }