+++ src/main/kotlin/app/revanced/api/configuration/routes/PatchesRoute.kt
@@ -1,11 +1,10 @@
 package app.revanced.api.configuration.routes
 
+import app.revanced.api.configuration.ApiAssetPublicKey
+import app.revanced.api.configuration.ApiRelease
+import app.revanced.api.configuration.ApiReleaseVersion
 import app.revanced.api.configuration.installCache
 import app.revanced.api.configuration.installNotarizedRoute
-import app.revanced.api.configuration.schema.APIAssetPublicKeys
-import app.revanced.api.configuration.schema.APIPatchesAsset
-import app.revanced.api.configuration.schema.APIRelease
-import app.revanced.api.configuration.schema.APIReleaseVersion
 import app.revanced.api.configuration.services.PatchesService
 import io.bkbn.kompendium.core.metadata.GetInfo
 import io.ktor.http.*
@@ -17,18 +16,9 @@ import kotlin.time.Duration.Companion.days
 import org.koin.ktor.ext.get as koinGet
 
 internal fun Route.patchesRoute() = route("patches") {
-    configure()
-
-    // TODO: Remove this deprecated route eventually.
-    route("latest") {
-        configure(deprecated = true)
-    }
-}
-
-private fun Route.configure(deprecated: Boolean = false) {
     val patchesService = koinGet<PatchesService>()
 
-    installPatchesRouteDocumentation(deprecated)
+    installPatchesRouteDocumentation()
 
     rateLimit(RateLimitName("weak")) {
         get {
@@ -36,7 +26,7 @@ private fun Route.configure(deprecated: Boolean = false) {
         }
 
         route("version") {
-            installPatchesVersionRouteDocumentation(deprecated)
+            installPatchesVersionRouteDocumentation()
 
             get {
                 call.respond(patchesService.latestVersion())
@@ -46,7 +36,7 @@ private fun Route.configure(deprecated: Boolean = false) {
 
     rateLimit(RateLimitName("strong")) {
         route("list") {
-            installPatchesListRouteDocumentation(deprecated)
+            installPatchesListRouteDocumentation()
 
             get {
                 call.respondBytes(ContentType.Application.Json) { patchesService.list() }
@@ -58,52 +48,49 @@ private fun Route.configure(deprecated: Boolean = false) {
         route("keys") {
             installCache(356.days)
 
-            installPatchesPublicKeyRouteDocumentation(deprecated)
+            installPatchesPublicKeyRouteDocumentation()
 
             get {
-                call.respond(patchesService.publicKeys())
+                call.respond(patchesService.publicKey())
             }
         }
     }
 }
 
-private fun Route.installPatchesRouteDocumentation(deprecated: Boolean) = installNotarizedRoute {
+private fun Route.installPatchesRouteDocumentation() = installNotarizedRoute {
     tags = setOf("Patches")
 
     get = GetInfo.builder {
-        if (deprecated) isDeprecated()
         description("Get the current patches release")
         summary("Get current patches release")
         response {
             description("The current patches release")
             mediaTypes("application/json")
             responseCode(HttpStatusCode.OK)
-            responseType<APIRelease<APIPatchesAsset>>()
+            responseType<ApiRelease>()
         }
     }
 }
 
-private fun Route.installPatchesVersionRouteDocumentation(deprecated: Boolean) = installNotarizedRoute {
+private fun Route.installPatchesVersionRouteDocumentation() = installNotarizedRoute {
     tags = setOf("Patches")
 
     get = GetInfo.builder {
-        if (deprecated) isDeprecated()
         description("Get the current patches release version")
         summary("Get current patches release version")
         response {
             description("The current patches release version")
             mediaTypes("application/json")
             responseCode(HttpStatusCode.OK)
-            responseType<APIReleaseVersion>()
+            responseType<ApiReleaseVersion>()
         }
     }
 }
 
-private fun Route.installPatchesListRouteDocumentation(deprecated: Boolean) = installNotarizedRoute {
+private fun Route.installPatchesListRouteDocumentation() = installNotarizedRoute {
     tags = setOf("Patches")
 
     get = GetInfo.builder {
-        if (deprecated) isDeprecated()
         description("Get the list of patches from the current patches release")
         summary("Get list of patches from current patches release")
         response {
@@ -115,18 +102,17 @@ private fun Route.installPatchesListRouteDocumentation(deprecated: Boolean) = in
     }
 }
 
-private fun Route.installPatchesPublicKeyRouteDocumentation(deprecated: Boolean) = installNotarizedRoute {
+private fun Route.installPatchesPublicKeyRouteDocumentation() = installNotarizedRoute {
     tags = setOf("Patches")
 
     get = GetInfo.builder {
-        if (deprecated) isDeprecated()
-        description("Get the public keys for verifying patches and integrations assets")
-        summary("Get patches and integrations public keys")
+        description("Get the public keys for verifying patches assets")
+        summary("Get patches public keys")
         response {
             description("The public keys")
             mediaTypes("application/json")
             responseCode(HttpStatusCode.OK)
-            responseType<APIAssetPublicKeys>()
+            responseType<ApiAssetPublicKey>()
         }
     }
 }