+++ src/main/kotlin/app/revanced/api/configuration/services/ApiService.kt
@@ -1,8 +1,9 @@
 package app.revanced.api.configuration.services
 
+import app.revanced.api.configuration.*
 import app.revanced.api.configuration.repository.BackendRepository
 import app.revanced.api.configuration.repository.ConfigurationRepository
-import app.revanced.api.configuration.schema.*
+import io.ktor.http.*
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
@@ -16,12 +17,16 @@ internal class ApiService(
     val about = configurationRepository.about
 
     suspend fun contributors() = withContext(Dispatchers.IO) {
-        configurationRepository.contributorsRepositoryNames.map {
+        configurationRepository.contributorsRepositoryNames.map { (repository, name) ->
             async {
                 APIContributable(
-                    it,
-                    backendRepository.contributors(configurationRepository.organization, it).map {
-                        APIContributor(it.name, it.avatarUrl, it.url, it.contributions)
+                    name,
+                    URLBuilder().apply {
+                        takeFrom(backendRepository.website)
+                        path(configurationRepository.organization, repository)
+                    }.buildString(),
+                    backendRepository.contributors(configurationRepository.organization, repository).map {
+                        ApiContributor(it.name, it.avatarUrl, it.url, it.contributions)
                     },
                 )
             }
@@ -29,13 +34,13 @@ internal class ApiService(
     }.awaitAll()
 
     suspend fun team() = backendRepository.members(configurationRepository.organization).map { member ->
-        APIMember(
+        ApiMember(
             member.name,
             member.avatarUrl,
             member.url,
             member.bio,
             if (member.gpgKeys.ids.isNotEmpty()) {
-                APIGpgKey(
+                ApiGpgKey(
                     // Must choose one of the GPG keys, because it does not make sense to have multiple GPG keys for the API.
                     member.gpgKeys.ids.first(),
                     member.gpgKeys.url,
@@ -47,6 +52,6 @@ internal class ApiService(
     }
 
     suspend fun rateLimit() = backendRepository.rateLimit()?.let {
-        APIRateLimit(it.limit, it.remaining, it.reset)
+        ApiRateLimit(it.limit, it.remaining, it.reset)
     }
 }