+++ src/main/kotlin/app/revanced/api/configuration/services/ManagerService.kt
@@ -1,38 +1,35 @@
 package app.revanced.api.configuration.services
 
+import app.revanced.api.configuration.ApiRelease
+import app.revanced.api.configuration.ApiReleaseVersion
 import app.revanced.api.configuration.repository.BackendRepository
 import app.revanced.api.configuration.repository.BackendRepository.BackendOrganization.BackendRepository.BackendRelease.Companion.first
 import app.revanced.api.configuration.repository.ConfigurationRepository
-import app.revanced.api.configuration.schema.*
 
 internal class ManagerService(
     private val backendRepository: BackendRepository,
     private val configurationRepository: ConfigurationRepository,
 ) {
-    suspend fun latestRelease(): APIRelease<APIManagerAsset> {
+    suspend fun latestRelease(): ApiRelease {
         val managerRelease = backendRepository.release(
             configurationRepository.organization,
             configurationRepository.manager.repository,
         )
 
-        val managerAsset = APIManagerAsset(
-            managerRelease.assets.first(configurationRepository.manager.assetRegex).downloadUrl,
-        )
-
-        return APIRelease(
+        return ApiRelease(
             managerRelease.tag,
             managerRelease.createdAt,
             managerRelease.releaseNote,
-            listOf(managerAsset),
+            managerRelease.assets.first(configurationRepository.manager.assetRegex).downloadUrl,
         )
     }
 
-    suspend fun latestVersion(): APIReleaseVersion {
+    suspend fun latestVersion(): ApiReleaseVersion {
         val managerRelease = backendRepository.release(
             configurationRepository.organization,
             configurationRepository.manager.repository,
         )
 
-        return APIReleaseVersion(managerRelease.tag)
+        return ApiReleaseVersion(managerRelease.tag)
     }
 }